+++ Changelog.md
@@ -5,6 +5,7 @@
 
 - Added the possibility to use "ilike" in all generic partial matching filters
 - Added the possibility to use a different field for the query than the id of the filter
+- Added the `NotNullTriState` filter type: it is very similar to `BoolTriState` except that it checks against null/not null on a given field
 - Changed the minimum Laravel version requirements to v10.48, v11.46.2 and v12.38 respectively
 - Added PHP 8.5 support
 - Added node v24 support